+++ crates/app/src/commands/graph/action.rs
@@ -1,10 +1,9 @@
-use crate::commands::graph::utils::{action_graph_repr, respond_to_request, setup_server};
+use crate::commands::graph::utils::{action_graph_repr, run_server};
 use crate::session::CliSession;
 use clap::Args;
 use moon_action_graph::RunRequirements;
 use moon_task::Target;
 use starbase::AppResult;
-use starbase_styles::color;
 use starbase_utils::json;
 use tracing::instrument;
 
@@ -38,9 +37,9 @@ pub async fn action_graph(session: CliSession, args: ActionGraphArgs) -> AppResu
         for target in targets {
             action_graph_builder.run_task_by_target(target, &requirements)?;
         }
-
-        // Show all targets and actions
-    } else {
+    }
+    // Show all targets and actions
+    else {
         for project in workspace_graph.get_projects()? {
             for task in workspace_graph.get_tasks_from_project(&project.id)? {
                 action_graph_builder.run_task(&project, &task, &requirements)?;
@@ -64,15 +63,7 @@ pub async fn action_graph(session: CliSession, args: ActionGraphArgs) -> AppResu
         return Ok(None);
     }
 
-    let (server, mut tera) = setup_server().await?;
-    let url = format!("http://{}", server.server_addr());
-    let _ = open::that(&url);
-
-    println!("Started server on {}", color::url(url));
-
-    for req in server.incoming_requests() {
-        respond_to_request(req, &mut tera, &graph_info, "Action graph".to_owned())?;
-    }
+    run_server("Action graph", graph_info).await?;
 
     Ok(None)
 }