+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Updated `moon task-graph`, `moon-project-graph`, and `moon action-graph` to request a port
+  dynamically, instead of defaulting to 8000.
+
+#### 🐞 Fixes
+
+- Fixed an issue where `moon *-graph` commands could not be CTRL+C'd.
+- Fixed an issue where Git based template locations would fail if the initial `git clone` failed.
+
 ## 1.32.6
 
 #### 🐞 Fixes